+++ adr/1_account.md
@@ -1,7 +1,7 @@
 # Account
 
 ## ER
-![](https://github.com/lightyeardev/backend-test-task/blob/main/adr/resource/1_account.svg?raw=true)
+![](https://github.com/bobsantos/backend-test-task/blob/main/adr/resource/1_account.svg?raw=true)
 
 ## Endpoints
