+++ adr/2_balance.md
@@ -1,0 +1,9 @@
+# Balance
+
+## ER
+![](https://github.com/bobsantos/backend-test-task/blob/main/adr/resource/2_balance.svg?raw=true)
+
+## Context
+
+Create a balance for a created account.
+
