+++ src/main/java/com/golightyear/backend/account/domain/Account.java
@@ -27,5 +27,4 @@ public class Account {
 
     @Builder.Default
     Instant lastModified = Instant.now();
-
 }
