+++ src/main/java/com/golightyear/backend/account/persistence/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.golightyear.backend.account;
+package com.golightyear.backend.account.persistence;
 
 import com.golightyear.backend.account.domain.Account;
 import com.golightyear.backend.account.domain.AccountId;
