+++ src/main/java/com/golightyear/backend/account/presentation/AccountResponse.java
@@ -1,5 +1,9 @@
-package com.golightyear.backend.account.domain;
+package com.golightyear.backend.account.presentation;
 
+import com.golightyear.backend.account.domain.Account;
+import com.golightyear.backend.account.domain.AccountId;
+import com.golightyear.backend.account.domain.AccountName;
+import com.golightyear.backend.account.domain.AccountState;
 import lombok.Value;
 
 import java.time.Instant;
